+++ example/lib/src/component_library/config/component_library.dart
@@ -15,6 +15,7 @@ import 'package:impaktfull_ui_example/src/component_library/items/calendar/calen
 import 'package:impaktfull_ui_example/src/component_library/items/card/card_library_item.dart';
 import 'package:impaktfull_ui_example/src/component_library/items/carrousel/carrousel_library_item.dart';
 import 'package:impaktfull_ui_example/src/component_library/items/chat/chat_library_item.dart';
+import 'package:impaktfull_ui_example/src/component_library/items/check_box_list_item/check_box_list_item_library_item.dart';
 import 'package:impaktfull_ui_example/src/component_library/items/checkbox/checkbox_library_item.dart';
 import 'package:impaktfull_ui_example/src/component_library/items/clamped_fractionally_sized_box/clamped_fractionally_sized_box_library_item.dart';
 import 'package:impaktfull_ui_example/src/component_library/items/cms_header/cms_header_library_item.dart';
@@ -112,6 +113,7 @@ class ComponentLibrary {
     const CarrouselLibraryItem(),
     const ChatLibraryItem(),
     const CheckboxLibraryItem(),
+    const CheckBoxListItemLibraryItem(),
     const ClampedFractionallySizedBoxLibraryItem(),
     const CmsHeaderLibraryItem(),
     const ColorPickerLibraryItem(),