+++ lib/src/components/check_box_list_item/check_box_list_item.describe.dart
@@ -1,0 +1,8 @@
+part of 'check_box_list_item.dart';
+
+String _describeInstance(
+    BuildContext context, ImpaktfullUiCheckBoxListItem instance) {
+  final descriptor = ComponentDescriptor();
+  descriptor.add('theme', instance.theme);
+  return descriptor.describe();
+}